+++ client/src/App.tsx
@@ -3,23 +3,7 @@ import { RouterProvider } from 'react-router-dom';
 import { AudioContextProvider } from './contexts/AudioContext';
 import AudioPlayer from './layouts/components/AudioPlayer/AudioPlayer';
 import { router } from './routes/router';
-import { createContext, useContext, useState } from 'react';
-
-type TAppContext = {
-  mobileOpen:any;
-  setMobileOpen:any;
-}
-
-export const AppContext = createContext<any>(null);
-
-function AppContextProvider({children}: any){
-      const [isClosing, setIsClosing] = useState(false);
-      const [mobileOpen, setMobileOpen] = useState(false);
-
-  return <AppContext.Provider value={{mobileOpen, setMobileOpen}}>
-    {children}
-  </AppContext.Provider>
-}
+import { AppContextProvider } from './contexts/AppContext';
 
 function App() {
   return (
